move GetCourseDetails endpoint to CourseController
